feature: main.py with swager documentation checked
