--- server/app/models.py
@@ -1,1 +1,0 @@
-# Models
