--- server/requirements.txt
@@ -1,4 +1,0 @@
-fastapi
-uvicorn
-python-dotenv
-requests
